Return the original type for null RLE block
During Presto-to-Velox expression conversion, the current logic only keeps
the original type when a single null RLE block is a DECIMAL, which causes
a mismatched-type problem that often leads to query failures (see the added
query test for details). Therefore, it is needed to make sure the original
type is always assigned for single null RLE block.