[native pos] Add serializedBytes runtime stats to PartitionAndSerialize
We are seeing Presto-on-Spark native producing a lot more shuffle data(in bytes)
then Presto-on-Spark Java. We believe this is because native uses UnsafeRow
serialization format which is less compact than row-wise serde used in
Presto-on-Spark Java.

We also noticed that there is no easy way to tell how many bytes we shuffle.

Add 'serializedBytes' runtime stat to PartitionAndSerialize to track total number
of bytes we shuffle.