[native] Fix exchange client creation.
During exchange client creation, ip address and port is passed to PrestoExchangeSource
constructor. IP address is passed as a string, which is needed to be converted to
IPAddress object. In https://github.com/prestodb/presto/pull/20650, this string to
IPAddress object conversion was removed, which caused all queries to fail. The reason
is in [SocketAddress](https://github.com/facebook/folly/blob/main/folly/SocketAddress.h#L53-L60)
constructor, host's corresponding IP will be looked up if allowNameLookup is True. In this case
allowNameLookup is set to True. Now, the host string is itself ip address, so such look up fails. This
change reverts the client creation and also removes allowNameLookup setting to true, as its not needed,
and also expensive.