[native] Fix path to json in protocol template and update generated files
Fix updating the path to the json dependency in the presto protocol
template file. The resulting generated files (presto_protocol.h/.cpp)
are updated to reflect the new change.
The original change was made manually incorrectly by PR #20749 (same
changes as in the initial PR #20619).

In addition, a comment is updated to reflect the new path where
it was referencing the old path.