[native] Fix Timestamp construction for negative timestamps
In Velox, `Timestamp::seconds_` is the starting point of one second,
and `Timestamp::nanos_` is the offset from the starting point (in
nanoseconds). So when converting milliseconds to seconds, we should
always round down, including negative values. This means that it is
wrong to use division and modulo directly. For example, for `-400ms`,
it should be converted into `(-1, 6 * 10^8)` instead of
`(0, -4 * 10^8)`.

`velox::Timestamp::fromMillis()` utility function can be used to
achieve this.